+++ app/src/main/java/com/finance_tracker/finance_tracker/theme/AppColors.kt
@@ -10,4 +10,5 @@ object AppColors {
     val Teal700 = Color(0xFF018786)
     val Black = Color(0xFF000000)
     val White = Color(0xFFFFFFFF)
+    val Green = Color(0xFF00BC2D)
 }
