+++ app/src/main/res/drawable/ic_calendar.xml
@@ -1,11 +1,9 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="18dp"
-    android:height="18dp"
-    android:viewportWidth="18"
-    android:viewportHeight="18">
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
   <path
       android:fillColor="#FF000000"
-      android:pathData="M12.5,2.4h2.8a0.7,0.7 0,0 1,0.7 0.7V14.3a0.7,0.7 0,0 1,-0.7 0.7H2.7a0.7,0.7 0,0 1,-0.7 -0.7V3.1a0.7,0.7 0,0 1,0.7 -0.7H5.5V1H6.9V2.4h4.2V1h1.4ZM11.1,3.8H6.9V5.2H5.5V3.8H3.4V6.6H14.6V3.8H12.5V5.2H11.1ZM14.6,8H3.4v5.6H14.6Z"
-      android:strokeAlpha="1"
-      android:fillAlpha="1"/>
+      android:pathData="M17,3h4a1,1 0,0 1,1 1L22,20a1,1 0,0 1,-1 1L3,21a1,1 0,0 1,-1 -1L2,4A1,1 0,0 1,3 3L7,3L7,1L9,1L9,3h6L15,1h2ZM15,5L9,5L9,7L7,7L7,5L4,5L4,9L20,9L20,5L17,5L17,7L15,7ZM20,11L4,11v8L20,19Z"/>
 </vector>
